Started fixing darwing interface enumeration issue.
