+++ responder3/core/interfaces/NetworkInterface.py
@@ -1,0 +1,35 @@
+import json
+
+class NetworkInterface:
+	def __init__(self):
+		"""
+		Container object to describe a network interface
+		"""
+		self.ifname = None
+		self.ifindex = None #zone_indices in windows
+		self.addresses = []
+	
+	def to_dict(self):
+		t = {}
+		t['ifname'] = str(self.ifname)
+		t['ifindex'] = str(self.ifindex)
+		t['addresses'] = []
+		for addr in self.addresses:
+			t['addresses'].append(str(addr))
+		return t
+		
+	def to_json(self):
+		return json.dumps(self.to_dict())
+
+	def __repr__(self):
+		return str(self)
+		
+	def __str__(self):
+		t  = '== INTERFACE ==\r\n'
+		t += 'Name: %s\r\n' % self.ifname
+		t += 'ifindex: %s\r\n' % self.ifindex
+		for addr in self.addresses:
+			t += 'Address: %s\r\n' % str(addr)
+		
+		return t
+
