+++ responder3/core/interfaces/README.md
@@ -1,0 +1,15 @@
+Linux implementation thankyous:
+
+This code is based on the following implementation: https://gist.github.com/chadmiller/5157850#file-getnifs-py
+In the original code it is mentioned that it's actually a modified version of: https://gist.github.com/provegard/1536682
+Which was in fact based on the "getifaddrs.py" script from pydlnadms [http://code.google.com/p/pydlnadms/].
+
+Windows implementations thankyous:
+
+The windows version is based on the code from https://github.com/darkk
+https://github.com/darkk/tcp_shutter/blob/master/tcp_shutter.py
+
+OSX (MAC) implementation thankyous:
+
+Implementation was created for this project by EvilWan (https://github.com/evilwan)
+
