+++ responder3/core/interfaces/test.py
@@ -1,0 +1,7 @@
+#!/usr/bin/env python3
+
+from NetworkInterfaces import NetworkInterfaces
+
+res = NetworkInterfaces()
+print(res)
+
